feat: package dbc and cms data at build time
This is the beginning of the deprecation of the CMS.
It also allows the EIC to run without having to reach
out to the cloud after load to acquire data. Still need
to look into server worker to be able to startup the
EIC without any network connectivity at all.